requirements.txt

fastapi
uvicorn
imageio
pillow
numpy
scipy
